+++ frontend/cypress.config.ts
@@ -8,6 +8,5 @@ export default defineConfig({
     e2e: {
         baseUrl: 'http://127.0.0.1:4200',
         video: false,
-        responseTimeout: 60000,
     },
 });