+++ frontend/src/app/pages/exercises/exercise/shared/exercise-map/feature-managers/vehicle-feature-manager.ts
@@ -1,11 +1,11 @@
 import type { Store } from '@ngrx/store';
+import { normalZoom } from 'digital-fuesim-manv-shared';
 import type {
     UUID,
     Vehicle,
     // eslint-disable-next-line @typescript-eslint/no-shadow
     Element,
 } from 'digital-fuesim-manv-shared';
-import { normalZoom } from 'digital-fuesim-manv-shared';
 import type { Feature, MapBrowserEvent } from 'ol';
 import type Point from 'ol/geom/Point';
 import type { TranslateEvent } from 'ol/interaction/Translate';
@@ -14,6 +14,7 @@ import type { Subject } from 'rxjs';
 import type { ExerciseService } from 'src/app/core/exercise.service';
 import type { AppState } from 'src/app/state/app.state';
 import { selectVisibleVehicles } from 'src/app/state/application/selectors/shared.selectors';
+import { Fill, Stroke } from 'ol/style';
 import { VehiclePopupComponent } from '../shared/vehicle-popup/vehicle-popup.component';
 import type { OlMapInteractionsManager } from '../utility/ol-map-interactions-manager';
 import { PointGeometryHelper } from '../utility/point-geometry-helper';
@@ -21,6 +22,7 @@ import { ImagePopupHelper } from '../utility/popup-helper';
 import { ImageStyleHelper } from '../utility/style-helper/image-style-helper';
 import { NameStyleHelper } from '../utility/style-helper/name-style-helper';
 import type { PopupService } from '../utility/popup.service';
+import { CircleStyleHelper } from '../utility/style-helper/circle-style-helper';
 import { MoveableFeatureManager } from './moveable-feature-manager';
 
 export class VehicleFeatureManager extends MoveableFeatureManager<Vehicle> {
@@ -50,6 +52,26 @@ export class VehicleFeatureManager extends MoveableFeatureManager<Vehicle> {
     );
     private readonly popupHelper = new ImagePopupHelper(this.olMap, this.layer);
 
+    private readonly openPopupCircleStyleHelper = new CircleStyleHelper(
+        (feature) => ({
+            radius: Math.max(
+                (this.getElementFromFeature(feature) as Vehicle).image.height,
+                (this.getElementFromFeature(feature) as Vehicle).image.height *
+                    (this.getElementFromFeature(feature) as Vehicle).image
+                        .aspectRatio
+            ),
+            fill: new Fill({
+                color: '#00000000',
+            }),
+            stroke: new Stroke({
+                color: 'orange',
+                width: 10,
+            }),
+        }),
+        0.025,
+        (_) => [0, 0]
+    );
+
     constructor(
         olMap: OlMap,
         private readonly store: Store<AppState>,
@@ -68,10 +90,23 @@ export class VehicleFeatureManager extends MoveableFeatureManager<Vehicle> {
             new PointGeometryHelper(),
             1000
         );
-        this.layer.setStyle((feature, resolution) => [
-            this.nameStyleHelper.getStyle(feature as Feature, resolution),
-            this.imageStyleHelper.getStyle(feature as Feature, resolution),
-        ]);
+        this.layer.setStyle((feature, resolution) => {
+            const styles = [
+                this.nameStyleHelper.getStyle(feature as Feature, resolution),
+                this.imageStyleHelper.getStyle(feature as Feature, resolution),
+            ];
+            this.addMarking(
+                feature,
+                styles,
+                this.popupService,
+                this.store,
+                this.openPopupCircleStyleHelper.getStyle(
+                    feature as Feature,
+                    resolution
+                )
+            );
+            return styles;
+        });
     }
 
     public override onFeatureDrop(
@@ -116,11 +151,20 @@ export class VehicleFeatureManager extends MoveableFeatureManager<Vehicle> {
     ): void {
         super.onFeatureClicked(event, feature);
 
+        const vehicle = this.getElementFromFeature(feature) as Vehicle;
+
         this.popupService.openPopup(
             this.popupHelper.getPopupOptions(
                 VehiclePopupComponent,
                 feature,
                 [feature.getId() as UUID],
+                [
+                    ...Object.keys(vehicle.materialIds),
+                    ...Object.keys(vehicle.personnelIds),
+                    feature.getId() as UUID,
+                ],
+                [feature.getId() as UUID],
+                ['vehicle', 'personnel', 'material'],
                 {
                     vehicleId: feature.getId() as UUID,
                 }