+++ shared/src/models/radiogram/transfer-category-completed-radiogram.ts
@@ -1,0 +1,54 @@
+import { IsBoolean, IsUUID, ValidateNested } from 'class-validator';
+import { UUID, uuidValidationOptions } from '../../utils';
+import { IsLiteralUnion, IsValue } from '../../utils/validators';
+import { IsRadiogramStatus } from '../../utils/validators/is-radiogram-status';
+import { getCreate, PatientStatus, patientStatusAllowedValues } from '../utils';
+import type { Radiogram } from './radiogram';
+import { ExerciseRadiogramStatus } from './status/exercise-radiogram-status';
+import {
+    Scope as TransferProgressScope,
+    scopeAllowedValues,
+} from './utils/transfer-progress-scope';
+
+export class TransferCategoryCompletedRadiogram implements Radiogram {
+    @IsUUID(4, uuidValidationOptions)
+    readonly id: UUID;
+
+    @IsValue('transferCategoryCompletedRadiogram')
+    readonly type = 'transferCategoryCompletedRadiogram';
+
+    @IsUUID(4, uuidValidationOptions)
+    readonly simulatedRegionId: UUID;
+
+    /**
+     * @deprecated use the helpers from {@link radiogram-helpers.ts}
+     * or {@link radiogram-helpers-mutable.ts} instead
+     */
+    @IsRadiogramStatus()
+    @ValidateNested()
+    readonly status: ExerciseRadiogramStatus;
+
+    @IsBoolean()
+    readonly informationAvailable: boolean = false;
+
+    @IsLiteralUnion(patientStatusAllowedValues)
+    readonly completedCategory: PatientStatus = 'white';
+
+    @IsLiteralUnion(scopeAllowedValues)
+    readonly scope: TransferProgressScope = 'singleRegion';
+
+    /**
+     * @deprecated Use {@link create} instead
+     */
+    constructor(
+        id: UUID,
+        simulatedRegionId: UUID,
+        status: ExerciseRadiogramStatus
+    ) {
+        this.id = id;
+        this.simulatedRegionId = simulatedRegionId;
+        this.status = status;
+    }
+
+    static readonly create = getCreate(this);
+}