+++ shared/src/simulation/events/transfer-patients-in-specific-vehicle-request.ts
@@ -20,6 +20,9 @@ export class TransferPatientsInSpecificVehicleRequestEvent
     @IsUUID(4, uuidValidationOptions)
     readonly vehicleId: UUID;
 
+    @IsUUID(4, uuidValidationOptions)
+    readonly transferInitiatingRegionId?: UUID;
+
     @IsLiteralUnion(transferDestinationTypeAllowedValues)
     readonly transferDestinationType: TransferDestination;
 
@@ -33,10 +36,12 @@ export class TransferPatientsInSpecificVehicleRequestEvent
         patientIds: UUIDSet,
         vehicleId: UUID,
         transferDestinationType: TransferDestination,
-        transferDestinationId: UUID
+        transferDestinationId: UUID,
+        transferInitiatingRegionId?: UUID
     ) {
         this.patientIds = patientIds;
         this.vehicleId = vehicleId;
+        this.transferInitiatingRegionId = transferInitiatingRegionId;
         this.transferDestinationType = transferDestinationType;
         this.transferDestinationId = transferDestinationId;
     }