+++ shared/src/simulation/events/transfer-vehicles-request.ts
@@ -1,5 +1,10 @@
 import { IsOptional, IsString, IsUUID } from 'class-validator';
-import { getCreate } from '../../models/utils';
+import { Type } from 'class-transformer';
+import {
+    ExerciseOccupation,
+    getCreate,
+    occupationTypeOptions,
+} from '../../models/utils';
 import { IsLiteralUnion, IsValue } from '../../utils/validators';
 import { ResourceDescription } from '../../models/utils/resource-description';
 import { IsResourceDescription } from '../../utils/validators/is-resource-description';
@@ -17,11 +22,18 @@ export class TransferVehiclesRequestEvent implements SimulationEvent {
     @IsResourceDescription()
     readonly requestedVehicles: ResourceDescription;
 
+    @IsUUID(4, uuidValidationOptions)
+    readonly transferInitiatingRegionId?: UUID;
+
     @IsLiteralUnion(transferDestinationTypeAllowedValues)
     readonly transferDestinationType: TransferDestination;
 
     @IsUUID(4, uuidValidationOptions)
     readonly transferDestinationId: UUID;
+
+    @IsOptional()
+    @Type(...occupationTypeOptions)
+    readonly successorOccupation?: ExerciseOccupation;
 
     @IsOptional()
     @IsString()
@@ -34,12 +46,16 @@ export class TransferVehiclesRequestEvent implements SimulationEvent {
         requestedVehicles: ResourceDescription,
         transferDestinationType: TransferDestination,
         transferDestinationId: UUID,
-        key?: string
+        transferInitiatingRegionId?: UUID,
+        key?: string,
+        successorOccupation?: ExerciseOccupation
     ) {
         this.requestedVehicles = requestedVehicles;
+        this.transferInitiatingRegionId = transferInitiatingRegionId;
         this.transferDestinationType = transferDestinationType;
         this.transferDestinationId = transferDestinationId;
         this.key = key;
+        this.successorOccupation = successorOccupation;
     }
 
     static readonly create = getCreate(this);